Update and migrate project
Move build file to Kotlin.
Move to Kotlin.
Remove options to select trigger event, trigger only in post-build as
  specified by the action.
Add option to fail the build when the request fails.
Add more useful data that is being POSTed to the webhook.
Add secret for identification with the webhook.
Add X-Jenkins-Token and X-Jenkins-Event headers in webhook request.
Use Gson.
Bump version to 0.3.0-SNAPSHOT.